Faster and more memory efficient headers for HTTP, HTTP/2, STOMP and SPYD. Fixes #3600
Motivation:

I noticed that the headers implementation in Netty for HTTP/2 uses quite a lot of memory
and that also at least the performance of randomly accessing a header is quite poor. The main
concern however was memory usage, as profiling has shown that a DefaultHttp2Headers
not only use a lot of memory it also wastes a lot due to the underlying hashmaps having
to be resized potentially several times as new headers are being inserted.

This is tracked as issue #3600.

Modifications:

I started out doing some research regarding possible implementations and my first idea
was to simply implement the headers using two Object[], one for the names and one for
the values as well as an int[] to cache the hashcode. That way iteration and insertion
would be blazingly fast, while random access would be O(n). Surprisingly, this
implementation outperformed the current implementation even for random access up to a
size of roughly 32 header entries. However, after the IntObjectHashMap got a massive
performance improvement by @nmittler (#3627), random access performance was not
competitive anymore and so I looked for alternatives.

I was looking for a solution with decent memory usage, competitive performance and low
complexity. That's how I arrived at the current implementation.

I redesigned the DefaultHeaders to simply take a Map object in its constructor and
reimplemented the class using only the Map primitives. That way the implementation
is very concise and hopefully easy to understand and it allows each concrete headers
implementation to provide its own map or to even use a different headers implementation
for processing requests and writing responses i.e. incoming headers need to provide
fast random access while outgoing headers need fast insertion and fast iteration. The
new implementation can support this with hardly any code changes i.e. provide
DefaultHttpHeaders backed by a HashMap for incoming requests and provide ResponseHttpHeaders
for outgoing requests with a Map simply backed by an ArrayList for minimal memory usage
and blazing fast insertion and iteration. This idea is not part of this PR though. It also comes
with the advantage that if the Netty project decides to add a third party collections library
as a dependency, one can simply plug in one of those very fast and memory efficient map
implementations and get faster and smaller headers for free.

For now, I am using the JDK's TreeMap for HTTP and HTTP/2 default headers. While probably
quite controversial I believe it's a decent choice for both protocols:
 1) HTTP header names are case insensitive and the HttpHeaders interface allows them to be
    used with any CharSequence e.g. AsciiString and String. In order to support this,
    the current implementation uses a HashCodeGenerator and NameConverter interface,
    which besides being problematic regarding the equals and hashCode() contract, is
    also quite complex. A TreeMap doesn't need any of those as it can be passed a custom
    Comparator. For now, I am simply using the case insensitive comparator already implemented
    in AsciiString.
 2) HTTP/2 Headers actually need to be ordered, namely all reserved header names must come
    first when written on the wire. In order to support this, we currently do something like
    the following before passing the headers to HPACK:
        for each reservedHeader in reservedHeaders
           if headers.contains(reservedHeader)
               write(reservedHeader)
        for each header in headers:
           if not reservedHeaders.contains(header):
               write(header)
    So it would be nice if we could avoid this and simply ensure the required order within
    the DefaultHttp2Headers. A small Comparator passed to TreeMap does the job.
 3) When full, a HashMap needs to be resized. That means the two backing arrays have to be
    thrown out, a new array must be allocated and all keys have to be rehashed. TreeMap does
    not require any of this.
 4) It has decent performance and memory usage and it's already in the JDK, we don't need to implement or
    maintain it.

I also added new tests, fixed (and introduced) some bugs and cleaned up the comments.

Result:

- Significantly fewer lines of code in the implementation. While the total commit is still
  roughly 400 lines less, the actual implementation is a lot less. I just added some more
  tests and microbenchmarks.

- Overall performance is up. The current implementation should be significantly faster
  for insertion and retrieval. However, it is slower when it comes to iteration. There is simply
  no way a TreeMap can have the same iteration performance as a linked list (as used in the
  current headers implementation). That's totally fine though, because when looking at the
  benchmark results @ejona86 pointed out that the performance of the headers is completely
  dominated by insertion, that is insertion is so significantly faster in the new implementation
  that it does make up for several times the iteration speed. You can't iterate what you haven't
  inserted. I am demonstrating that in this spreadsheet [1]. (Actually, iteration performance is
  only down for HTTP, it's significantly improved for HTTP/2).

- Memory is down. The implementation with TreeMap uses on avg ~30% less memory. It also does not
  produce any garbage while being resized. In load tests for GRPC we have seen a memory reduction
  of up to 1.2KB per RPC. I summarized the memory improvements in this spreadsheet [1]. The data
  was generated by [2] using JOL.

- While it was my original intend to only improve the memory usage for HTTP/2, it should be similarly
  improved for HTTP, SPDY and STOMP as they all share a common implementation.

[1] https://docs.google.com/spreadsheets/d/1ck3RQklyzEcCLlyJoqDXPCWRGVUuS-ArZf0etSXLVDQ/edit#gid=0
[2] https://gist.github.com/buchgr/4458a8bdb51dd58c82b4

bla